WindowsFormsCalc/WindowsFormsCalc/Form1.cs
using System;
using System.Collections.Generic;
using System.ComponentModel;
using System.Data;
using System.Drawing;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using System.Windows.Forms;

namespace WindowsFormsCalc
{
    public partial class Form1 : Form
    {
        public static int funct_calc(int a, int b, char op)
        {
            switch (op)
            {
                case '+':
                    return a + b;
                case '-':
                    return a - b;
                case '/':
                    return a / b;
                case '*':
                    return a * b;
            }
            throw new ArgumentException("Illegal operator");
        }

        public Form1()
        {
            InitializeComponent();
        }

        private void button1_Click(object sender, EventArgs e)
        {
            int res = 0;
            int a = Convert.ToInt32(axtx.Text);
            int b = Convert.ToInt32(btxt.Text);
            char op = Convert.ToChar(this.op.Text);
            res = funct_calc(a, b, op);
            restxt.Text = Convert.ToString(res);
        }

        private void Form1_Load(object sender, EventArgs e)
        {

        }

        private void textBox1_TextChanged(object sender, EventArgs e)
        {

        }
    }
}
